conanfile.py: Use QTDIR instead of QT_PATH
There's an established environment variable for this though, QTDIR.
This environment variable is also set by Qt Creator by default.
So the Conan recipes should accept this environment variable instead.

This requires a change in the matching conan profile files
where the path to Qt installation is injected to env.

Task-number: QTBUG-90583
Change-Id: I14408ba952dd5de652a52db26e85b8ca01dcfc88
Reviewed-by: Kai Koehne <kai.koehne@qt.io>