conanfile.py: Use QTDIR instead of QT_PATH
There's an established environment variable for this though, QTDIR.
This environment variable is also set by Qt Creator by default.
So the Conan recipes should accept this environment variable instead.

This requires a change in the matching conan profile files
where the path to Qt installation is injected to env.

Task-number: QTBUG-90583
Change-Id: I3b1675a555583a6738fdf413abc0524b1f906ca2
Reviewed-by: Jesus Fernandez <jsfdez@gmail.com>