conanfile.py: Use QTDIR instead of QT_PATH
There's an established environment variable for this though, QTDIR.
This environment variable is also set by Qt Creator by default.
So the Conan recipes should accept this environment variable instead.

This requires a change in the matching conan profile files
where the path to Qt installation is injected to env.

Pick-to: 6.1
Task-number: QTBUG-90583
Change-Id: I46b62069d2f3368100aa97fd4eb50e98960830d7
Reviewed-by: Antti Kokko <antti.kokko@qt.io>