conanfile.py: Use QTDIR instead of QT_PATH
There's an established environment variable for this though, QTDIR.
This environment variable is also set by Qt Creator by default.
So the Conan recipes should accept this environment variable instead.

This requires a change in the matching conan profile files
where the path to Qt installation is injected to env.

Task-number: QTBUG-90583
Pick-to: 6.1
Change-Id: I128cc1390f0a8525ee4490393cb63338bf511c16
Reviewed-by: Simo Fält <simo.falt@qt.io>