Add conanfile.py to support builds with Conan
Support building with Conan against Qt6 base binary package.
Default the 'shared' option to the value of the Qt build.

The build recipe expects to locate QTDIR from env.
The 'qt-cmake(.bat)' is used for the build.

'CMAKE_TOOLCHAIN_FILE' is used to pick the toolchain
file for cross compilations.

For Android builds 'ANDROID_SDK_ROOT' and
'ANDROID_NDK_ROOT' need to be passed from env.

Pick-to: 6.1
Change-Id: Iaba057aef9d72b57aa919445c017b57b5a713ece
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>