Update conan support
This patch updates the conanfile.py to support the full Qt build using
conan. qtbase is now listed as a dependency, and the build recipe uses
the qtbase's 'qt-configure-module' script directly.

Use shared recipe functionality from qt-conan-common package.

This change is also required for QtPositioning module, as it has an
optional dependency on QtSerialPort.

Pick-to: 6.2
Task-number: QTBUG-90800
Change-Id: I54f9f717f8fec98f9617385c8533b116ad0e9d1c
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Toni Saario <toni.saario@qt.io>