Conan: Inherit recipe class from QtLeafModule for common functionality
The qt-conan-common package implements a base class for Qt leaf
module recipes. The build steps in leaf modules are mostly identical
so it makes sense to put those in the base class.

Dependencies are read by the base class from the 'dependencies.yaml'
which is the same file the CI system uses.

Pick-to: 6.2
Task-number: QTBUG-94866
Change-Id: I0e581eec9e547961712e6d704f2c818688ab2b8e
Reviewed-by: Toni Saario <toni.saario@qt.io>