Conan: Inherit recipe class from QtLeafModule for common functionality
The qt-conan-common package implements a base class for Qt leaf
module recipes. The build steps in leaf modules are mostly identical
so it makes sense to put those in the base class.

Dependencies are read by the base class from the 'dependencies.yaml'
which is the same file the CI system uses.

Change-Id: Ib5cc5ab4c5b387a5ca1ee2587c9e785c151cde34
Reviewed-by: Toni Saario <toni.saario@qt.io>