Conan: Add a recipe to pack build profiles in a Conan package
We want to distribute the used build profiles as well. For this we
can put the profiles in a separate Conan package that can be
installed easily.

This is a "header only" recipe which merely packs the profiles inside
the Conan package which the consumers can install with "conan install"
command:

  $conan install qtbuildprofiles/6.2.0@qt/everywhere --build=always

Result:

  ./windows-x86_64-mingw
  ./linux-x86_64-gcc
  ./macos-ios-universal-clang
  ./...

Pick-to: 6.2
Change-Id: I745a089b4e855c006e1a5f2f0995a73abbcebf2f
Reviewed-by: Toni Saario <toni.saario@qt.io>